Implemented database updates and login verification

+++ main.js
@@ -1,4 +1,5 @@
-var database = [];
+// Initialize the database array with existing users from local storage, if any.
+let database = JSON.parse(localStorage.getItem("users")) || [];
 
 function signUp(e) {
   e.preventDefault();
@@ -16,9 +17,39 @@ function signUp(e) {
     confirmPassword: confirmPassword,
   };
 
-  database.push(newUser);
-  localStorage.setItem("users", JSON.stringify(database));
+  if (password !== confirmPassword) {
+    alert(`Password does not match`);
+  } else {
+    database.push(newUser);
 
-  console.log(database);
-  window.location.href = "./signIn.html";
+    // Save the updated database back to local storage.
+    localStorage.setItem("users", JSON.stringify(database));
+    window.location.href = "./signIn.html";
+  }
+}
+
+function login(e) {
+  e.preventDefault();
+  let email = document.getElementById("loginEmail").value;
+  let password = document.getElementById("loginPassword").value;
+
+  let users = JSON.parse(localStorage.getItem("users"));
+  if (users) {
+    let user = users.find(
+      (user) => user.email === email && user.password === password
+    );
+    if (user) {
+      alert(`Login Successful`);
+      //   clearAllInput();
+    } else {
+      alert(`Invalid Login credentials`);
+    }
+  } else {
+    alert(`No user found`);
+  }
+
+  //   function clearAllInput() {
+  //     email.value = "";
+  //     password.value = "";
+  //   }
 }
